// src/app/api/catalog/model/showEpisodeResponseDto.ts
/**
 * CinePik Catalog API
 * The CinePik Catalog microservice.
 *
 * The version of the OpenAPI document: 0.3.2
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SourceResponseDto } from './sourceResponseDto';


export interface ShowEpisodeResponseDto { 
    /**
     * Episode identifier.
     */
    id: number;
    /**
     * Show identifier.
     */
    show_id: number;
    /**
     * Episode number.
     */
    episode_number: number;
    /**
     * Episode season number.
     */
    season_number: number;
    /**
     * Url for the thumbnail image.
     */
    thumbnail_path: string;
    /**
     * Episode first aired date.
     */
    first_aired?: string;
    /**
     * Episode title.
     */
    title: string;
    /**
     * Whether the episode is currently available.
     */
    availability: boolean;
    /**
     * Show sources.
     */
    sources: Array<SourceResponseDto>;
}