// src/app/api/catalog/model/similarMovieDetailResponseDto.ts
/**
 * CinePik Catalog API
 * The CinePik Catalog microservice.
 *
 * The version of the OpenAPI document: 0.3.2
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SimilarMovieDetailResponseDto { 
    /**
     * Movie identifier.
     */
    id: number;
    /**
     * Url for the poster image.
     */
    poster_path: string;
    /**
     * Movie title.
     */
    title: string;
    /**
     * Url for the backdrop image.
     */
    backdrop_path: string;
}